Agent 3 First Commit
Still some stuff to be worked out (sometimes it fails to solve a solvable maze, and the trajectory length seems really low at times, so I'll keep working on that). Other than that, I'll try to optimize a little bit more as well. It's really slow at the higher maze dimensions.