Agent 4 Framework
Please ask questions if you have them. This does not have to be what Agent 4 turns out to be like, but it's just my idea. If you want to try to implement your own way, make a separate folder and go for it. I'll try to explain my though process in the chat.